+++ aiecs/tools/office_tool/call_api.py
@@ -1,8 +1,0 @@
-# deprecated: use gateway.call_api
-from aiecs.tools.office_tool.gateway.call_api import (  # noqa: F401
-    CALL_API_DESCRIPTION,
-    OFFICE_CALL_API_TOOL,
-    office_call_api,
-)
-
-__all__ = ["CALL_API_DESCRIPTION", "OFFICE_CALL_API_TOOL", "office_call_api"]