+++ aiecs/tools/office_tool/docbuilder_script.py
@@ -1,2 +1,0 @@
-# deprecated: use core.docbuilder_script
-from aiecs.tools.office_tool.core.docbuilder_script import *  # noqa: F403,F401