+++ aiecs/tools/office_tool/execute_builder.py
@@ -1,7 +1,0 @@
-# deprecated: use gateway.execute_builder
-from aiecs.tools.office_tool.gateway.execute_builder import (  # noqa: F401
-    OFFICE_EXECUTE_BUILDER_TOOL,
-    office_execute_builder,
-)
-
-__all__ = ["OFFICE_EXECUTE_BUILDER_TOOL", "office_execute_builder"]