+++ aiecs/tools/office_tool/object_fetch.py
@@ -1,2 +1,0 @@
-# deprecated: use core.storage.object_fetch
-from aiecs.tools.office_tool.core.storage.object_fetch import *  # noqa: F403,F401