+++ aiecs/tools/office_tool/source_resolver.py
@@ -1,2 +1,0 @@
-# deprecated: use core.source
-from aiecs.tools.office_tool.core.source import *  # noqa: F403,F401