+++ aiecs/tools/office_tool/storage_paths.py
@@ -1,2 +1,0 @@
-# deprecated: use core.storage.paths
-from aiecs.tools.office_tool.core.storage.paths import *  # noqa: F403,F401